Databases are done(check comments)
I separated file editing(database) and modifying objects in the program(databaseGetter). All the data now initializes whenever you need it, and will save to the database if you type "end" to end the program. If you end the program prematurely, it won't save tho
DatabaseGetters are pretty intuitive. They are singleton classes though, which could change in the future.