Modified for StudentUser class, InstructorUser class and UserManager class and created the UserFactory class and IUser Interface
The phrase 1 feedback from TA was to implement factory design method. In order to implement factory design method to increase the extensibility, feasibility and independently of our program, I modified our codes for StudentUser class, InstructorUser class, and UserManager class and created the UserFactory class and IUser Interface as well as removed the user entity.